project/local-ai-server/requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow==10.0.1
flask==3.0.0
flask-cors==4.0.0
accelerate==0.24.0
bitsandbytes==0.41.0